Fix chat controller and user controller(user profile )

+++ backend/unithis/src/main/java/com/unithis/mapper/ChatRoomMapper.java
@@ -11,7 +11,7 @@ import com.unithis.model.ChatRoomRequest;
 @Mapper
 public interface ChatRoomMapper {
 	public int createChatRoom(ChatRoom newRoom);
-	public List<ChatRoom> getJoinedRoomList(int userId);
-	public ChatRoom checkExistChatRoom(@Param("user1") int user1, @Param("user2") int user2);
+	public List<ChatRoom> getJoinedRoomList(long userId);
+	public ChatRoom checkExistChatRoom(@Param("user1") long user1, @Param("user2") long user2);
 	public int updateReadTime(ChatRoomRequest chatroomReq);
 }
