+++ backend/unithis/src/main/java/com/unithis/service/ChatRoomService.java
@@ -28,7 +28,7 @@ public class ChatRoomService implements IChatRoomService {
 	}
 
 	@Override
-	public ChatRoom checkExistChatRoom(int user1, int user2) {
+	public ChatRoom checkExistChatRoom(long user1, long user2) {
 		return chatroomMapper.checkExistChatRoom(user1, user2);
 	}
